+++ src/test/java/com/senzing/listener/communication/rabbitmq/MockRabbitMQChannel.java
@@ -94,7 +94,7 @@ public class MockRabbitMQChannel implements Channel {
     /**
      * Gets the count of messages in the queue.
      */
-    public int getMessageCount() throws SQLException {
+    public long getMessageCount() throws SQLException {
         try (Connection conn = connectionPool.acquire()) {
             return sqlClient.getMessageCount(conn);
         }
